Enable recovery from errors thrown by handlers.
Currently, when a handler throws, any remaining handlers for the event
are not called. In Atom, when buggy packages throw the remaining
packages do not get called, and typically the default processing
also stops. As a result a buggy package can cause other well behaved
packages to stop working correctly.

This is particularly troublesome when a handler for TextBuffer
will-save throws as this prevents the user from being able to save
their file.

With this change, notification of handlers which throw can be had by
subscribing to Emitter.onHandlerThrow. Emmitter.setEnsureHandlerInvoke(true)
will guarantee that all handlers will get called, even if one of them
throws. Emitter.isEnsureHandlerInvoke() returns the current state of the
ensure handler Invoke bit.